+++ facebook_auth_platform_interface/lib/src/facebook_auth_implementation.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 import 'package:flutter/services.dart';
-import 'package:flutter/foundation.dart' show visibleForTesting, defaultTargetPlatform, TargetPlatform;
+import 'package:flutter/foundation.dart'
+    show visibleForTesting, defaultTargetPlatform, TargetPlatform;
 
 import 'facebook_auth_plaftorm.dart';
 import 'facebook_permissions.dart';
@@ -14,7 +15,8 @@ class FacebookAuthPlatformImplementation extends FacebookAuthPlatform {
   bool get isAndroid => defaultTargetPlatform == TargetPlatform.android;
 
   @visibleForTesting
-  MethodChannel channel = const MethodChannel('app.meedu/flutter_facebook_auth');
+  MethodChannel channel =
+      const MethodChannel('app.meedu/flutter_facebook_auth');
 
   /// make a login request using the facebook SDK
   ///
@@ -73,7 +75,9 @@ class FacebookAuthPlatformImplementation extends FacebookAuthPlatform {
     final result = await channel.invokeMethod("getUserData", {
       "fields": fields,
     });
-    return isAndroid ? jsonDecode(result) : Map<String, dynamic>.from(result); //null  or dynamic data
+    return isAndroid
+        ? jsonDecode(result)
+        : Map<String, dynamic>.from(result); //null  or dynamic data
   }
 
   /// Sign Out from Facebook